Move AbstractOpenWeatherMapSensor to sensor.py
