+++ homeassistant/components/openweathermap/sensor.py
@@ -1,13 +1,22 @@
 """Support for the OpenWeatherMap (OWM) service."""
-from .abstract_owm_sensor import AbstractOpenWeatherMapSensor
+from homeassistant.components.sensor import SensorEntity
+from homeassistant.const import ATTR_ATTRIBUTION
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
+
 from .const import (
     ATTR_API_FORECAST,
+    ATTRIBUTION,
+    DEFAULT_NAME,
     DOMAIN,
     ENTRY_NAME,
     ENTRY_WEATHER_COORDINATOR,
     FORECAST_MONITORED_CONDITIONS,
     FORECAST_SENSOR_TYPES,
+    MANUFACTURER,
     MONITORED_CONDITIONS,
+    SENSOR_DEVICE_CLASS,
+    SENSOR_NAME,
+    SENSOR_UNIT,
     WEATHER_SENSOR_TYPES,
 )
 from .weather_update_coordinator import WeatherUpdateCoordinator
@@ -48,6 +57,88 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
         )
 
     async_add_entities(entities)
+
+
+class AbstractOpenWeatherMapSensor(SensorEntity):
+    """Abstract class for an OpenWeatherMap sensor."""
+
+    def __init__(
+        self,
+        name,
+        unique_id,
+        sensor_type,
+        sensor_configuration,
+        coordinator: DataUpdateCoordinator,
+    ):
+        """Initialize the sensor."""
+        self._name = name
+        self._unique_id = unique_id
+        self._sensor_type = sensor_type
+        self._sensor_name = sensor_configuration[SENSOR_NAME]
+        self._unit_of_measurement = sensor_configuration.get(SENSOR_UNIT)
+        self._device_class = sensor_configuration.get(SENSOR_DEVICE_CLASS)
+        self._coordinator = coordinator
+
+    @property
+    def name(self):
+        """Return the name of the sensor."""
+        return f"{self._name} {self._sensor_name}"
+
+    @property
+    def unique_id(self):
+        """Return a unique_id for this entity."""
+        return self._unique_id
+
+    @property
+    def device_info(self):
+        """Return the device info."""
+        split_unique_id = self._unique_id.split("-")
+        return {
+            "identifiers": {(DOMAIN, f"{split_unique_id[0]}-{split_unique_id[1]}")},
+            "name": DEFAULT_NAME,
+            "manufacturer": MANUFACTURER,
+            "entry_type": "service",
+        }
+
+    @property
+    def should_poll(self):
+        """Return the polling requirement of the entity."""
+        return False
+
+    @property
+    def attribution(self):
+        """Return the attribution."""
+        return ATTRIBUTION
+
+    @property
+    def device_class(self):
+        """Return the device_class."""
+        return self._device_class
+
+    @property
+    def native_unit_of_measurement(self):
+        """Return the unit of measurement of this entity, if any."""
+        return self._unit_of_measurement
+
+    @property
+    def extra_state_attributes(self):
+        """Return the state attributes."""
+        return {ATTR_ATTRIBUTION: ATTRIBUTION}
+
+    @property
+    def available(self):
+        """Return True if entity is available."""
+        return self._coordinator.last_update_success
+
+    async def async_added_to_hass(self):
+        """Connect to dispatcher listening for entity data notifications."""
+        self.async_on_remove(
+            self._coordinator.async_add_listener(self.async_write_ha_state)
+        )
+
+    async def async_update(self):
+        """Get the latest data from OWM and updates the states."""
+        await self._coordinator.async_request_refresh()
 
 
 class OpenWeatherMapSensor(AbstractOpenWeatherMapSensor):
